ai-agents/v3/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
chromadb
pydantic
numpy
python-multipart
openai
tiktoken